+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The changelog for `Superwall`. Also see the [releases](https://github.com/superwall/react-native-superwall/releases) on GitHub.
 
+## 1.2.3
+
+### Enhancements
+
+- Adds `Superwall.shared.dismiss()` to be able to dismiss a paywall programmatically.
+- Upgrades Android SDK to 1.2.1. [View Android SDK release notes](https://github.com/superwall-me/Superwall-Android/releases/tag/1.1.7)
+
 ## 1.2.2
 
 ### Enhancements